Moved the inner controller to a separate file.
changed the inner controller to be the sole writer of the wheelspeed metric
